Removes content providers
